Refactor `test_geom`
